refact: rename anomalycard to routecard
